+++ services/blackbox/backend/app/abilities/main.ts
@@ -1,5 +1,0 @@
-import { Bouncer } from '@adonisjs/bouncer'
-
-export const editPost = Bouncer.ability(() => {
-  return true
-})