+++ services/blackbox/backend/app/controllers/auth/session_controller.ts
@@ -1,7 +1,0 @@
-import { routeController } from '@folie/castle'
-
-export default routeController({
-  handle: async ({ ctx: { session: auth } }) => {
-    return { session: auth.user.$serialize() }
-  },
-})