+++ services/blackbox/backend/app/controllers/ping_controller.ts
@@ -1,7 +1,0 @@
-import { routeController } from '@folie/castle'
-
-export default routeController({
-  handle: async () => {
-    return { message: 'pong' }
-  },
-})