+++ services/blackbox/backend/app/controllers/vault/key/show_controller.ts
@@ -1,7 +1,0 @@
-import { routeController } from '@folie/castle'
-
-export default routeController({
-  handle: async ({ ctx }) => {
-    return { key: ctx.session.user.key }
-  },
-})