+++ services/blackbox/backend/config/captcha.ts
@@ -1,7 +1,0 @@
-import env from '#start/env'
-import { Secret } from '@adonisjs/core/helpers'
-import { CaptchaService } from '@folie/castle/service/captcha_service'
-
-export const captcha = new CaptchaService({
-  privateKey: new Secret(env.get('CAPTCHA_PRIVATE_KEY')),
-})