+++ services/blackbox/frontend/src/components/query_loader.tsx
@@ -1,5 +1,0 @@
-import { QueryLoaderProps, QueryLoader } from "@folie/cobalt/components";
-
-export const LocalQueryLoader = <OUT,>(props: QueryLoaderProps<OUT>) => {
-  return <QueryLoader {...props} />;
-};