+++ services/blackbox/frontend/src/configs/index.ts
@@ -1,8 +1,0 @@
-import type { CobaltConfig as CC } from "@folie/cobalt/types";
-
-export type CobaltKeys = "formId";
-
-export const CobaltConfig: CC = {
-  theme: "dark",
-  forceTheme: true,
-};