+++ services/blackbox/frontend/src/configs/routes.ts
@@ -1,5 +1,0 @@
-import { routes as backendRoutes } from "@folie/service-formation-backend/blueprint";
-
-export const routes = {
-  ...backendRoutes,
-};