+++ services/blackbox/frontend/src/lib/helpers/get_object_value.ts
@@ -1,3 +1,0 @@
-export const getObjectValue = (obj: Record<string, unknown>, key: string) => {
-  return obj[key];
-};