+++ services/blackbox/frontend/src/lib/hooks/use_window.ts
@@ -1,3 +1,0 @@
-export const useWindow = () => {
-  return typeof window !== "undefined" ? window : null;
-};