+++ packages/blueprint/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @folie/blueprint
 
+## 0.0.5
+
+### Patch Changes
+
+- Work in progress
+
 ## 0.0.4
 
 ### Patch Changes